Add PyTorch modules from original repo and https://github.com/CCSI-Toolset/MGN
Need to clean up a lot and modify for actual weather forecasting, but gives a good base to start with. Seems like the PyTorch code does most of what is needed for weather forecasting.

The main things left would be get the graph of the world, a new encoder for it, new decoder for the model back to the lat/lon grid.